# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderlust adventure travel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Achievement {
  id: string;
  name: string;
  description: string;
  icon: string;
  unlocked: boolean;
}

export interface User {
  id: string;
  name: string;
  email: string;
  avatarUrl?: string | null;
  completedTripsCount: number;
  favoritesCount: number;
  achievements: Achievement[];
}

export interface SessionResponse {
  user: User | null;
}

export interface LoginBody {
  email: string;
  name: string;
}

export interface Country {
  code: string;
  name: string;
  flag: string;
  region: string;
  activityCount: number;
  heroImage: string;
}

export interface EmergencyInfo {
  police: string;
  ambulance: string;
  fire: string;
  touristPolice?: string | null;
  embassyHotline?: string | null;
}

export interface Activity {
  id: string;
  name: string;
  type: string;
  countryCode: string;
  countryName: string;
  countryFlag: string;
  city: string;
  difficulty: string;
  durationDays: number;
  estimatedCost: number;
  rating: number;
  reviewCount: number;
  heroImage: string;
  shortDescription: string;
  weather: string;
}

export interface CountryDetail {
  code: string;
  name: string;
  flag: string;
  region: string;
  description: string;
  heroImage: string;
  currency: string;
  bestSeason: string;
  emergency: EmergencyInfo;
  activities: Activity[];
}

export interface ActivityDetail {
  id: string;
  name: string;
  type: string;
  countryCode: string;
  countryName: string;
  countryFlag: string;
  city: string;
  difficulty: string;
  durationDays: number;
  estimatedCost: number;
  rating: number;
  reviewCount: number;
  heroImage: string;
  gallery: string[];
  description: string;
  bestTimeToVisit: string;
  weather: string;
  latitude: number;
  longitude: number;
  requiredGear: string[];
  highlights: string[];
  isFavorited: boolean;
}

export interface TripPoint {
  lat: number;
  lng: number;
  timestamp: number;
}

export interface Trip {
  id: string;
  title: string;
  activityId?: string | null;
  activityName?: string | null;
  countryCode?: string | null;
  startedAt: string;
  endedAt: string;
  distanceKm: number;
  durationSec: number;
  avgSpeedKmh: number;
  notes?: string | null;
  path: TripPoint[];
}

export interface CreateTripBody {
  title: string;
  activityId?: string | null;
  countryCode?: string | null;
  startedAt: string;
  endedAt: string;
  distanceKm: number;
  durationSec: number;
  avgSpeedKmh: number;
  notes?: string | null;
  path: TripPoint[];
}

export interface Review {
  id: string;
  activityId: string;
  userName: string;
  userAvatar?: string | null;
  rating: number;
  title: string;
  body: string;
  images: string[];
  createdAt: string;
}

export interface CreateReviewBody {
  rating: number;
  title: string;
  body: string;
  images?: string[];
}

export interface PackingItem {
  id: string;
  name: string;
  essential: boolean;
}

export interface PackingCategory {
  category: string;
  items: PackingItem[];
}

export interface PackingList {
  activityId: string;
  activityName: string;
  categories: PackingCategory[];
}

export interface DashboardSummary {
  totalAdventures: number;
  totalDistanceKm: number;
  countriesVisited: number;
  favoritesCount: number;
  recentTrips: Trip[];
  recommended: Activity[];
}

export interface TrendingDestination {
  countryCode: string;
  countryName: string;
  flag: string;
  heroImage: string;
  activityCount: number;
  avgRating: number;
}

export type ListActivitiesParams = {
  search?: string;
  country?: string;
  type?: string;
  difficulty?: string;
  minBudget?: number;
  maxBudget?: number;
  maxDurationDays?: number;
  weather?: string;
};

export type ListReviewsParams = {
  sort?: ListReviewsSort;
};

export type ListReviewsSort =
  (typeof ListReviewsSort)[keyof typeof ListReviewsSort];

export const ListReviewsSort = {
  recent: "recent",
  top: "top",
  low: "low",
} as const;