# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wanderlust adventure travel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /countries/{code}:
    get:
      operationId: getCountry
      tags: [countries]
      summary: Get country details with emergency info
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Country
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountryDetail"
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: Search and filter adventure activities
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: country
          in: query
          required: false
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
        - name: difficulty
          in: query
          required: false
          schema:
            type: string
        - name: minBudget
          in: query
          required: false
          schema:
            type: number
        - name: maxBudget
          in: query
          required: false
          schema:
            type: number
        - name: maxDurationDays
          in: query
          required: false
          schema:
            type: integer
        - name: weather
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Filtered list of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /activities/{id}:
    get:
      operationId: getActivity
      tags: [activities]
      summary: Get full activity details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActivityDetail"
...
  /activities/{id}/reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for an activity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum: [recent, top, low]
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "200":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /favorites/{activityId}:
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Favorite an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Added
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Unfavorite an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /packing/{activityId}:
    get:
      operationId: getPackingList
      tags: [packing]
      summary: Generate a smart packing list for an activity
      parameters:
        - name: activityId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Packing list grouped by category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackingList"
...
components:
  schemas:
...
    EmergencyInfo:
      type: object
      properties:
        police:
          type: string
        ambulance:
          type: string
        fire:
          type: string
        touristPolice:
          type: string
          nullable: true
        embassyHotline:
          type: string
          nullable: true
      required: [police, ambulance, fire]
    CountryDetail:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        flag:
          type: string
        region:
          type: string
        description:
          type: string
        heroImage:
          type: string
        currency:
          type: string
        bestSeason:
          type: string
        emergency:
          $ref: "#/components/schemas/EmergencyInfo"
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
      required: [code, name, flag, region, description, heroImage, currency, bestSeason, emergency, activities]
    Activity:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        countryCode:
          type: string
        countryName:
          type: string
        countryFlag:
          type: string
        city:
          type: string
        difficulty:
          type: string
        durationDays:
          type: integer
        estimatedCost:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        heroImage:
          type: string
        shortDescription:
          type: string
        weather:
          type: string
      required:
        [id, name, type, countryCode, countryName, countryFlag, city, difficulty, durationDays, estimatedCost, rating, reviewCount, heroImage, shortDescription, weather]
    ActivityDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        countryCode:
          type: string
        countryName:
          type: string
        countryFlag:
          type: string
        city:
          type: string
        difficulty:
          type: string
        durationDays:
          type: integer
        estimatedCost:
          type: number
        rating:
          type: number
        reviewCount:
          type: integer
        heroImage:
          type: string
        gallery:
          type: array
          items:
            type: string
        description:
          type: string
        bestTimeToVisit:
          type: string
        weather:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        requiredGear:
          type: array
          items:
            type: string
        highlights:
          type: array
          items:
            type: string
        isFavorited:
          type: boolean
      required:
        [id, name, type, countryCode, countryName, countryFlag, city, difficulty, durationDays, estimatedCost, rating, reviewCount, heroImage, gallery, description, bestTimeToVisit, weather, latitude, longitude, requiredGear, highlights, isFavorited]
...
    Review:
      type: object
      properties:
        id:
          type: string
        activityId:
          type: string
        userName:
          type: string
        userAvatar:
          type: string
          nullable: true
        rating:
          type: integer
        title:
          type: string
        body:
          type: string
        images:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, activityId, userName, rating, title, body, images, createdAt]
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
        title:
          type: string
        body:
          type: string
        images:
          type: array
          items:
            type: string
      required: [rating, title, body]
    PackingItem:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        essential:
          type: boolean
      required: [id, name, essential]
    PackingCategory:
      type: object
      properties:
        category:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PackingItem"
      required: [category, items]
    PackingList:
      type: object
      properties:
        activityId:
          type: string
        activityName:
          type: string
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PackingCategory"
      required: [activityId, activityName, categories]
...