# lib/api-zod/src/generated/types/emergencyInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wanderlust adventure travel API
 * OpenAPI spec version: 0.1.0
 */

export interface EmergencyInfo {
  police: string;
  ambulance: string;
  fire: string;
  touristPolice?: string | null;
  embassyHotline?: string | null;
}